Restructure several files and functions
