+++ benchmarks/simulate_sequence_fisp.cpp
@@ -3,8 +3,10 @@
 #include <random>
 
 #include "common.hpp"
-#include "core/context.h"
-#include "simulate/sequence.h"
+#include "compas/core/complex_type.h"
+#include "compas/core/context.h"
+#include "compas/sequences/fisp.h"
+#include "compas/simulate/fisp.h"
 
 using namespace compas;
