+++ include/compas/simulate/pssfp.h
@@ -1,9 +1,8 @@
 #include "compas/parameters/tissue.h"
-#include "compas/sequences/fisp.h"
 #include "compas/sequences/pssfp.h"
+#include "compas/sequences/pssfp_view.h"
 
 namespace compas {
-
 Array<cfloat, 2> simulate_magnetization(
     const CudaContext& context,
     TissueParameters parameters,
@@ -14,16 +13,4 @@ void simulate_magnetization_kernel(
     cuda_view_mut<cfloat, 2> echos,
     TissueParametersView parameters,
     pSSFPSequenceView sequence);
-
-Array<cfloat, 2> simulate_magnetization(
-    const CudaContext& context,
-    TissueParameters parameters,
-    FISPSequence sequence);
-
-void simulate_magnetization_kernel(
-    const kmm::CudaDevice& context,
-    cuda_view_mut<cfloat, 2> echos,
-    TissueParametersView parameters,
-    FISPSequenceView sequence);
-
 }  // namespace compas
