+++ julia-bindings/src/bindings.cpp
@@ -1,16 +1,16 @@
+#include "compas/jacobian/hermitian.h"
+#include "compas/jacobian/product.h"
+#include "compas/parameters/tissue.h"
+#include "compas/simulate/derivative.h"
+#include "compas/simulate/fisp.h"
+#include "compas/simulate/pssfp.h"
+#include "compas/trajectories/cartesian.h"
+#include "compas/trajectories/phase_encoding.h"
+#include "compas/trajectories/signal.h"
+#include "compas/trajectories/spiral.h"
+#include "compas/trajectories/trajectory.h"
+#include "compas/utils/residual.h"
 #include "constants.h"
-#include "core/context.h"
-#include "jacobian/product.h"
-#include "parameters/tissue.h"
-#include "sequences/pssfp.h"
-#include "simulate/derivative.h"
-#include "simulate/phase_encoding.h"
-#include "simulate/residual.h"
-#include "simulate/sequence.h"
-#include "simulate/signal.h"
-#include "trajectories/cartesian.h"
-#include "trajectories/spiral.h"
-#include "trajectories/trajectory.h"
 
 // Alias for complex float.
 using cfloat = compas::complex_type<float>;
