+++ src/simulate/pssfp_kernels.cuh
@@ -3,9 +3,9 @@
 #include <cooperative_groups.h>
 #include <cooperative_groups/reduce.h>
 
+#include "../operators/isochromat.cuh"
 #include "compas/core/complex_type.h"
 #include "compas/sequences/pssfp_view.h"
-#include "../operators/isochromat.cuh"
 
 namespace compas {
 namespace kernels {
